requirements.txt
# Requirements for Math-Noise Project (Training & Evaluation)
# Optimized for remote Slurm clusters with NVIDIA A100 GPUs

# Core DL Frameworks
torch>=2.4.0
transformers>=4.44.0
datasets>=2.20.0
accelerate>=0.33.0

# SFT (Supervised Fine-Tuning) & LoRA
# Verl is compatible with older versions of trl where AutoModelForCausalLMWithValueHead is available
trl==0.9.6
peft>=0.12.0

# RL (Verl / PPO / GRPO)
# Note: Verl often needs to be installed from source or specific wheels
# These are the supporting packages for Verl
hydra-core>=1.3.2
omegaconf>=2.3.0
ray>=2.30.0
pandas>=2.0.0
pyarrow>=15.0.0

# Efficient Inference & Rollout
vllm==0.6.3.post1

# Monitoring & Progress
wandb>=0.17.0
tqdm>=4.66.0
requests>=2.32.0

# Numerical & Utilities
numpy>=1.24.0
scipy>=1.10.0
